+++ api/src/parser/atom_test.rs
@@ -704,26 +704,4 @@ mod atom_test {
         let atom = Atom::TheNumberPi;
         assert_eq!(&atom.to_string(), "[pi]")
     }
-
-    #[cfg(feature = "with_serde")]
-    mod with_serde {
-        use crate::parser::Atom;
-        use serde_json;
-
-        #[test]
-        fn validate_serialization() {
-            let j = serde_json::to_string(&Atom::BushelUS)
-                .expect("Couldn't convert Atom to JSON String");
-
-            assert_eq!("\"BushelUS\"", j);
-        }
-
-        #[test]
-        fn validate_deserialization() {
-            let k =
-                serde_json::from_str("\"BushelUS\"").expect("Couldn't convert JSON String to Atom");
-
-            assert_eq!(Atom::BushelUS, k);
-        }
-    }
 }