+++ api/src/unit.rs
@@ -15,7 +15,7 @@ mod term_reducing;
 pub mod to_reduced;
 pub mod ucum_unit;
 
-#[cfg(feature = "with_serde")]
+#[cfg(feature = "serde")]
 mod serde;
 
 use crate::parser::Term;